Add a toggle to sort the move list by asc or desc order
Challenge #4 in the ReactJS tic-tac-toe tutorial.

Renamed stepNumber to currentStep for better intuition.

Removed the Go To Game Start button at the start of the moves list because it complicated sorting; instead I move it to a new button somewhere else.

Added stepNumber to a history element to make it easier to still highlight the current move in a reversed move list.